🚚Move CategorizedKeyExtract to its own module/file
