+++ unittests/test_json_serialization.py
@@ -9,12 +9,8 @@ import pytest  # type:ignore[import]
 from lark import Token, Tree
 from marshmallow import Schema, ValidationError
 
-from ahbicht.content_evaluation.content_evaluation_result import (
-    CategorizedKeyExtract,
-    CategorizedKeyExtractSchema,
-    ContentEvaluationResult,
-    ContentEvaluationResultSchema,
-)
+from ahbicht.content_evaluation.categorized_key_extract import CategorizedKeyExtract, CategorizedKeyExtractSchema
+from ahbicht.content_evaluation.content_evaluation_result import ContentEvaluationResult, ContentEvaluationResultSchema
 from ahbicht.edifact import EdifactFormat
 from ahbicht.evaluation_results import (
     AhbExpressionEvaluationResult,
